requirements.txt
langchain
langchain-ollama
langchain-qdrant
qdrant-client
psycopg2-binary
pandas
PyPDF2
python-dotenv